+++ CHANGES.rst
@@ -56,6 +56,9 @@ pipeline
 stpipe
 ------
 
+- Limit reference file prefetch to the first "science" exptype
+  when a pipeline has an association as input. [#5031]
+
 - Remove further sloper references. [#4989]